# swagger/swagger.yml
basePath: /api/v1
definitions:
...
  inventory.ProductInput:
    properties:
      product_name:
        type: string
      sku:
        type: string
      unit_price:
        type: number
    required:
      - product_name
      - sku
      - unit_price
    type: object
  inventory.ProductOutput:
    properties:
      id:
        type: integer
      product_name:
        type: string
      sku:
        type: string
      unit_price:
        type: number
    type: object
...
  models.Customer:
    properties:
      address:
        type: string
      credit_balance:
        type: integer
      customer_id:
        type: string
      customer_name:
        type: string
      customer_type:
        type: string
      email:
        type: string
      image:
        type: string
      phone:
        type: string
    type: object
...
  sales.CreateCustomerRequest:
    properties:
      address:
        type: string
      credit_balance:
        type: integer
      customer_name:
        type: string
      customer_type:
        type: string
      email:
        type: string
      image:
        type: string
      phone:
        type: string
    required:
      - address
      - customer_name
      - customer_type
    type: object
  sales.UpdateCustomerRequest:
    properties:
      address:
        type: string
      credit_balance:
        type: integer
      customer_name:
        type: string
      customer_type:
        type: string
      email:
        type: string
      image:
        type: string
      phone:
        type: string
    required:
      - address
      - customer_name
      - customer_type
    type: object
host: localhost:8080
info:
  contact:
    email: hendri@kewr-digital.com
    name: Hendri Sulistya
  description: This is the API documentation for Kewr-ERP.
  title: Kewr-ERP API
  version: "1.0"
paths:
  /customers:
    get:
      description: Retrieve all customers from the database
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/models.Customer"
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
        - BearerAuth: []
      summary: Retrieve all customers
      tags:
        - Customers
    post:
      consumes:
        - application/json
      description: Create a new customer with the provided details
      parameters:
        - description: Customer details
          in: body
          name: customer
          required: true
          schema:
            $ref: "#/definitions/sales.CreateCustomerRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Customer"
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
        - BearerAuth: []
      summary: Create a new customer
      tags:
        - Customers
  /customers/{id}:
    delete:
      description: Delete a customer from the database by its ID
      parameters:
        - description: Customer ID
          in: path
          name: id
          required: true
          type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
        - BearerAuth: []
      summary: Delete a customer
      tags:
        - Customers
    get:
      description: Retrieve a customer from the database by its ID
      parameters:
        - description: Customer ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Customer"
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
        - BearerAuth: []
      summary: Retrieve a customer by ID
      tags:
        - Customers
    put:
      consumes:
        - application/json
      description: Update the details of an existing customer
      parameters:
        - description: Customer ID
          in: path
          name: id
          required: true
          type: string
        - description: Updated customer details
          in: body
          name: customer
          required: true
          schema:
            $ref: "#/definitions/sales.UpdateCustomerRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Customer"
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
        - BearerAuth: []
      summary: Update an existing customer
      tags:
        - Customers
...
  /products:
    get:
      description: Retrieve all products from the database
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/inventory.ProductOutput"
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
        - BearerAuth: []
      summary: Get all products
      tags:
        - Products
    post:
      consumes:
        - application/json
      description: Add a new product to the database
      parameters:
        - description: Product
          in: body
          name: product
          required: true
          schema:
            $ref: "#/definitions/inventory.ProductInput"
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/inventory.ProductOutput"
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
        - BearerAuth: []
      summary: Create a new product
      tags:
        - Products
  /products/{id}:
    delete:
      description: Delete a product from the database by its ID
      parameters:
        - description: Product ID
          in: path
          name: id
          required: true
          type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
        - BearerAuth: []
      summary: Delete a product
      tags:
        - Products
    get:
      description: Retrieve a product from the database by its ID
      parameters:
        - description: Product ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/inventory.ProductOutput"
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
        - BearerAuth: []
      summary: Get a product by ID
      tags:
        - Products
    put:
      consumes:
        - application/json
      description: Update a product in the database
      parameters:
        - description: Product ID
          in: path
          name: id
          required: true
          type: integer
        - description: Product
          in: body
          name: product
          required: true
          schema:
            $ref: "#/definitions/inventory.ProductInput"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/inventory.ProductOutput"
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
        - BearerAuth: []
      summary: Update a product
      tags:
        - Products
...
schemes:
  - http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"